pin Poison dependency to >= 1.4 and < 2
Prior to this commit, the Poison dependency was pinned to anything in
the 1.4.x series; however, given that Poison follows semantic
versioning, we should pin to 1.x (where x >= 4) instead. This will allow
compatability for dependency resolution when bundled with other projects
that may require a newer version of Poison, such as Phoenix.